Update the input code to handle views
In particular, weston_pointer and weston_touch both have a view as their
focus instead of a surface while weston_keyboard, on the other hand, still
has a surface as its focus.  This is because a surface has no visual
location while a view does.

Another thing worth noting here that will become a standard in future
commits is that, when a surface with a single view needs to be storred
(such as the pointer sprite), only the view is storred and not the surface.
We can do this because the view.surface pointer is guaranteed to be valid
and it solves the problem of having to find the view in the list of views.