Remove support for text cursor position based output zooming
The entire concept of a text cursor position requires a concept of a
"primary view".  While this could be implemented in a shell, the current
implementation crosses the shell/compositor line in strange ways.  For now,
it is simplest to just remove support for this.  If we decide we want it
in the future we can reimplement it then in a more sane way.