Update the xwayland interface to handle views
Because we have to inform the client every time it gets moved on screen, we
need to know what view to consider "primary" and report to the client.  To
do this, a get_primary_view vfunc is added to the shell interface that
takes a shell sruface and returns the view that should be considered
"primary".  The window manager will then listen to this view for changes
and update the X client accordingly.