Update the shell plugin to work with views.
This commit updates the shell plugin to work with views.  Most of it is
pretty solid, but there are still some issues to work out.  First, is that
how to handle transient surfaces isn't obvious.  I think the current
heuristic is pretty good, but it could use clerification.

Second, input panels are broken.  How to get these working with views in a
sane way is not obvious and will probably require more api changes.